requirements.txt
Django==4.2.25
djangorestframework==3.16.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.12.1
